Force to use int32

+++ laserchicken/test_spatial_selections.py
@@ -64,7 +64,7 @@ class TestSpatialSelectionWKT(unittest.TestCase):
             pc_out = points_in_polygon_wkt(pc_in, "POLYGON(( 243590.0 572110.0, 243640.0 572160.0, 243700.0 572110.0, 243640.0 572060.0, 243590.0 572110.0 ))")
             x = pc_out[point]['x']['data']
             y = pc_out[point]['y']['data']
-            df_out = pd.DataFrame({'x':x, 'y':y})
+            df_out = pd.DataFrame({'x':x, 'y':y}, dtype=np.int32)
             df = pd.read_csv("testdata/ahn2_polygon.out", sep=',', header=0, index_col=0, dtype=np.int32)
             assert(pd.DataFrame.equals(df, df_out))
         except ValueError as err:
@@ -150,7 +150,7 @@ class TestSpatialSelectionWKTFile(unittest.TestCase):
             pc_out = points_in_polygon_wkt_file(pc_in, "testdata/ahn2_geometries_wkt/ahn2_polygon.wkt")
             x = pc_out[point]['x']['data']
             y = pc_out[point]['y']['data']
-            df_out = pd.DataFrame({'x':x, 'y':y})
+            df_out = pd.DataFrame({'x':x, 'y':y}, dtype=np.int32)
             df = pd.read_csv("testdata/ahn2_polygon.out", sep=',', header=0, index_col=0, dtype=np.int32)
             assert(pd.DataFrame.equals(df, df_out))
         except ValueError as err:
@@ -232,7 +232,7 @@ class TestSpatialSelectionSHPFile(unittest.TestCase):
             pc_out = points_in_polygon_shp_file(pc_in, "testdata/ahn2_polygon_shp/ahn2_polygon.shp")
             x = pc_out[point]['x']['data']
             y = pc_out[point]['y']['data']
-            df_out = pd.DataFrame({'x': x, 'y': y})
+            df_out = pd.DataFrame({'x': x, 'y': y}, dtype=np.int32)
             df = pd.read_csv("testdata/ahn2_polygon.out", sep=',', header=0, index_col=0, dtype=np.int32)
             assert (pd.DataFrame.equals(df, df_out))
         except ValueError as err:
